// client/src/types/LiveCoinWatchData.ts
export interface LiveCoinWatchData {
  age: number;
  allTimeHighUSD: number;
  categories: string[];
  circulatingSupply: number;
  code: string;
  color: string;
  exchanges: number;
  history: any[];
  links: any;
  markets: number;
  maxSupply: any;
  name: string;
  pairs: number;
  png32: string;
  png64: string;
  rank: number;
  totalSupply: number;
  web32: string;
  web64: string;
}